test: Use fabs instead of fabsf for the tolerance comparison
In c4eb49ff15406 the probing was changed to use doubles but this was
missed.

+++ vkrunner/vr-test.c
@@ -654,7 +654,7 @@ compare_pixels(const double *color1,
                int n_components)
 {
         for (int p = 0; p < n_components; ++p)
-                if (fabsf(color1[p] - color2[p]) > tolerance[p])
+                if (fabs(color1[p] - color2[p]) > tolerance[p])
                         return false;
         return true;
 }
